Move old request and response to a separate file
